Move the j_blues properties from radiative_properties to a separate j_blues module.
